+++ package.json
@@ -1,52 +1,23 @@
 {
-  "name": "latex2js",
+  "name": "latex2js-workspace",
   "version": "0.0.1",
   "publishConfig": {
     "access": "restricted"
   },
   "private": true,
   "scripts": {
-    "commit": "git-cz",
-    "build": "lerna run build",
-    "bootstrap": "lerna bootstrap --use-workspaces",
-    "build:serve": "bin/build-all.sh && pnpm run serve",
-    "serve": "serve website/latex2js.com/static",
-    "test:watch": "jest --watch",
-    "test": "jest",
-    "type-check": "tsc --noEmit"
-  },
-  "commitlint": {
-    "extends": [
-      "@commitlint/config-conventional"
-    ]
-  },
-  "config": {
-    "commitizen": {
-      "path": "cz-conventional-changelog"
-    }
+    "build": "lerna run build"
   },
   "devDependencies": {
-    "@commitlint/cli": "^18.0.0",
-    "@commitlint/config-conventional": "^18.0.0",
+    "copyfiles": "^2.4.1",
+    "prettier": "^3.0.0",
+    "rimraf": "^5.0.5",
+    "ts-jest": "^29.1.0",
     "@types/jest": "^29.5.0",
     "@types/node": "^20.0.0",
-    "commitizen": "^4.3.0",
-    "commitlint": "^18.0.0",
-    "cz-conventional-changelog": "^3.3.0",
-    "husky": "^8.0.0",
     "jest": "^29.7.0",
     "lerna": "^7.4.0",
-    "prettier": "^3.0.0",
-    "stylelint": "^15.11.0",
-    "stylelint-config-recommended": "^13.0.0",
-    "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-processor-styled-components": "^1.10.0",
     "typescript": "^5.2.0"
-  },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
-    }
   },
   "workspaces": [
     "packages/*",
