+++ packages/html5/package.json
@@ -24,28 +24,15 @@
     "bundle": "browserify dist/index.js --s LaTeX2HTML5 -o dist/latex2html5.bundle.js",
     "test": "jest",
     "test:watch": "jest --watch",
-    "test:debug": "node --inspect node_modules/.bin/jest --runInBand",
-    "type-check": "tsc --noEmit"
+    "test:debug": "node --inspect node_modules/.bin/jest --runInBand"
   },
   "bugs": {
     "url": "https://github.com/pyramation/latex2js/issues"
   },
-  "devDependencies": {
-    "@types/d3": "^7.4.0",
-    "@types/jest": "^29.5.0",
-    "browserify": "^16.5.0",
-    "copyfiles": "^2.4.1",
-    "jest": "^29.7.0",
-    "prettier": "^3.0.0",
-    "rimraf": "^5.0.5",
-    "ts-jest": "^29.1.0",
-    "typescript": "^5.2.0"
-  },
   "dependencies": {
-    "@latex2js/macros": "workspace:*",
     "mathjaxjs": "workspace:*",
+    "latex2js": "workspace:*",
     "@latex2js/pstricks": "workspace:*",
-    "d3": "^7.8.0",
-    "latex2js": "workspace:*"
+    "@latex2js/macros": "workspace:*"
   }
 }
