--- packages/mathjax/CHANGELOG.md
@@ -1,44 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [2.1.0](https://github.com/pyramation/latex2js/compare/latex2js-mathjax@2.0.0...latex2js-mathjax@2.1.0) (2020-06-12)
-
-**Note:** Version bump only for package latex2js-mathjax
-
-
-
-
-
-# 2.0.0 (2020-06-12)
-
-
-
-# 1.1.0 (2018-03-08)
-
-
-
-## 1.0.5 (2018-03-07)
-
-
-
-## 0.1.10 (2018-03-07)
-
-
-
-## 0.0.30 (2018-03-07)
-
-
-
-## 0.0.29 (2018-03-07)
-
-
-
-## 0.0.28 (2018-03-07)
-
-
-
-## 0.0.27 (2018-03-07)
-
-**Note:** Version bump only for package latex2js-mathjax
