--- packages/mathjax/README.md
@@ -1,1 +1,0 @@
-# latex2js-mathjax
