--- packages/mathjax/package.json
@@ -1,64 +1,0 @@
-{
-  "name": "@latex2js/mathjax",
-  "version": "2.1.0",
-  "description": "MathJax utilities for LaTeX2JS",
-  "author": "Dan Lynch <pyramation@gmail.com>",
-  "homepage": "https://github.com/pyramation/latex2js/tree/master/packages/mathjax#readme",
-  "license": "SEE LICENSE IN LICENSE",
-  "main": "index.js",
-  "module": "esm/index.js",
-  "types": "index.d.ts",
-  "publishConfig": {
-    "access": "public",
-    "directory": "dist"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/pyramation/latex2js"
-  },
-  "keywords": [
-    "latex",
-    "latex2html",
-    "latex2js",
-    "math",
-    "mathematics",
-    "mathjax"
-  ],
-  "scripts": {
-    "copy": "copyfiles -f ../../LICENSE README.md package.json dist",
-    "clean": "rimraf dist",
-    "build": "pnpm run clean && tsc && tsc -p tsconfig.esm.json && pnpm run copy",
-    "build:dev": "pnpm run clean && tsc --declarationMap && tsc -p tsconfig.esm.json && pnpm run copy",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:debug": "node --inspect node_modules/.bin/jest --runInBand",
-    "type-check": "tsc --noEmit"
-  },
-  "bugs": {
-    "url": "https://github.com/pyramation/latex2js/issues"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.0",
-    "@types/node": "^20.0.0",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.0",
-    "typescript": "^5.2.0",
-    "copyfiles": "^2.4.1",
-    "rimraf": "^5.0.5"
-  },
-  "dependencies": {
-    "load-script": "^1.0.0"
-  },
-  "peerDependencies": {
-    "react": ">=16.8.0",
-    "next": ">=12.0.0"
-  },
-  "peerDependenciesMeta": {
-    "react": {
-      "optional": true
-    },
-    "next": {
-      "optional": true
-    }
-  }
-}
