--- packages/mathjax/src/load-script.d.ts
@@ -1,4 +1,0 @@
-declare module 'load-script' {
-  function loadScript(src: string, callback?: (error?: Error) => void): void;
-  export = loadScript;
-}
