--- packages/mathjax/tsconfig.esm.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist/esm",
-    "module": "es2022",
-    "rootDir": "src/",
-    "declaration": false
-  }
-}
