+++ packages/mathjaxjs-react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@latex2js/mathjax-react",
+  "name": "mathjaxjs-react",
   "version": "2.1.0",
   "description": "React-specific MathJax provider component",
   "author": "Dan Lynch <pyramation@gmail.com>",
@@ -29,8 +29,7 @@
     "build:dev": "pnpm run clean && tsc --declarationMap && tsc -p tsconfig.esm.json && pnpm run copy",
     "test": "jest",
     "test:watch": "jest --watch",
-    "test:debug": "node --inspect node_modules/.bin/jest --runInBand",
-    "type-check": "tsc --noEmit"
+    "test:debug": "node --inspect node_modules/.bin/jest --runInBand"
   },
   "bugs": {
     "url": "https://github.com/pyramation/latex2js/issues"
@@ -43,9 +42,6 @@
   },
   "devDependencies": {
     "@types/react": "^18.0.0",
-    "react": "^18.0.0",
-    "typescript": "^5.2.0",
-    "copyfiles": "^2.4.1",
-    "rimraf": "^5.0.5"
+    "react": "^18.0.0"
   }
 }
