+++ packages/mathjaxjs-react/src/components/MathJaxProvider.tsx
