+++ packages/mathjaxjs-react/src/index.ts
