+++ packages/mathjaxjs-react/tsconfig.esm.json
