+++ packages/mathjaxjs-react/tsconfig.json
