+++ packages/react/package.json
@@ -39,29 +39,18 @@
     "url": "https://github.com/pyramation/latex2js/issues"
   },
   "devDependencies": {
-    "@types/d3": "^7.4.0",
-    "@types/jest": "^29.5.0",
     "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.2.0",
-    "jest": "^29.7.0",
-    "prettier": "^3.0.0",
-    "regenerator-runtime": "^0.13.5",
-    "stylelint": "^15.11.0",
-    "stylelint-config-recommended": "^13.0.0",
-    "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-processor-styled-components": "^1.10.0",
-    "typescript": "^5.2.0",
-    "copyfiles": "^2.4.1",
-    "rimraf": "^5.0.5"
+    "@types/react-dom": "^18.2.0"
   },
   "dependencies": {
-    "d3": "^7.8.0",
-    "latex2js": "workspace:*",
-    "@latex2js/macros": "workspace:*",
-    "mathjaxjs": "workspace:*",
-    "@latex2js/pstricks": "workspace:*",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "mathjaxjs": "workspace:*",
+    "mathjaxjs-react": "workspace:*",
+    "latex2js": "workspace:*",
+    "@latex2js/utils": "workspace:*",
+    "@latex2js/pstricks": "workspace:*",
+    "@latex2js/macros": "workspace:*"
   },
   "peerDependencies": {
     "react": ">=16.8.0",
