--- packages/settings/tsconfig.build.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "declaration": true,
-    "declarationMap": true,
-    "module": "commonjs",
-    "target": "es2017"
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist", "main", "module"]
-}
