+++ packages/utils/package.json
@@ -31,19 +31,9 @@
     "build:dev": "pnpm run clean && tsc --declarationMap && tsc -p tsconfig.esm.json && pnpm run copy",
     "test": "jest",
     "test:watch": "jest --watch",
-    "test:debug": "node --inspect node_modules/.bin/jest --runInBand",
-    "type-check": "tsc --noEmit"
+    "test:debug": "node --inspect node_modules/.bin/jest --runInBand"
   },
   "bugs": {
     "url": "https://github.com/pyramation/latex2js/issues"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.0",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.0",
-    "typescript": "^5.2.0",
-    "copyfiles": "^2.4.1",
-    "rimraf": "^5.0.5"
-  },
-  "dependencies": {}
+  }
 }
